+++ apps/api/prisma/migrations/20260614014703_add_user_avatar_generation/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "core"."users" ADD COLUMN     "avatarGeneration" INTEGER NOT NULL DEFAULT 0;