+++ apps/api/src/infrastructure/redis/index.ts
@@ -1,4 +1,5 @@
 export { REDIS_CLIENT } from './redis.constants'
 export { RedisModule } from './redis.module'
 export { type AppRedisClient, RedisConnectionService } from './redis-connection.service'
-export { RedisLockService } from './redis-lock.service'
+export { type BlockingAcquireOptions, RedisLockService } from './redis-lock.service'
+export { LockUnavailableError, type MutexOptions, RedisMutexService } from './redis-mutex.service'